Move sourcee code to Sources/MagicTimer directory.
