+++ Sources/MagicTimer/Int+Extension.swift
