+++ Sources/MagicTimer/MGBackgroundCalculator.swift
