+++ Sources/MagicTimer/MGCounter.swift
