+++ Sources/MagicTimer/MGLogable.swift
