+++ Sources/MagicTimer/MGObservable.swift
