+++ Sources/MagicTimer/MGStandardTimerFormatter.swift
