+++ Sources/MagicTimer/MGStateManager.swift
