+++ Sources/MagicTimer/MGTimerExecutive.swift
