+++ Sources/MagicTimer/MagicTimerState.swift
