+++ Sources/MagicTimer/MagicTimerView.swift
