+++ Sources/MagicTimer/MagicTimerViewDelegate.swift
