+++ Sources/MagicTimer/TimeInterval+Extesnion.swift
